create database extended flexibility
now there won't be an error if one sends a command to create tables, but
they already exist! Each table is checked seperately (by name). 

Changed class names.

Added a method to sqlUtilities, to handle connections that are not in
the pool (like the one used to create the table).

maybe more...